requirements.txt
requests==2.32.3
ical==9.0.3
python-dateutil==2.9.0
